+++ src/Board.h
@@ -296,7 +296,7 @@ private:
 
   // state and timing
 
-  unsigned long configPhaseEnd; // for offering config mode
+  unsigned long configPhaseEnd; // millis when current config mode (boardstate) is over, next mode
   unsigned long connectPhaseEnd; // for waiting on net connection
   unsigned long _captiveEnd; // terminate/reset captive portal mode after 5 minutes.
   void _newState(BoardState newState);